fix: simple restructuring of the directories
