+++ graphs/depthfirstsearch/depthfirstsearch.go
@@ -1,6 +1,4 @@
-package graphs
-
-import "fmt"
+package depthfirstsearch
 
 func getIdx(target int, nodes []int) int {
 	for i := 0; i < len(nodes); i++ {
@@ -47,26 +45,26 @@ func dfs(start, end int, nodes []int, edges [][]bool) ([]int, bool) {
 	return nil, false
 }
 
-func main() {
-	nodes := []int{
-		1, 2, 3, 4, 5, 6,
-	}
-	/*
-		sample graph
-		①-②
-		|  |
-		③-④-⑤-⑥
-	*/
-	edges := [][]bool{
-		{false, true, true, false, false, false},
-		{true, false, false, true, false, false},
-		{true, false, false, true, false, false},
-		{false, true, true, false, true, false},
-		{false, false, false, true, false, true},
-		{false, false, false, false, true, false},
-	}
-	start := 1
-	end := 6
-	route, _ := dfs(start, end, nodes, edges)
-	fmt.Println(route)
-}
+// func main() {
+// 	nodes := []int{
+// 		1, 2, 3, 4, 5, 6,
+// 	}
+// 	/*
+// 		sample graph
+// 		①-②
+// 		|  |
+// 		③-④-⑤-⑥
+// 	*/
+// 	edges := [][]bool{
+// 		{false, true, true, false, false, false},
+// 		{true, false, false, true, false, false},
+// 		{true, false, false, true, false, false},
+// 		{false, true, true, false, true, false},
+// 		{false, false, false, true, false, true},
+// 		{false, false, false, false, true, false},
+// 	}
+// 	start := 1
+// 	end := 6
+// 	route, _ := dfs(start, end, nodes, edges)
+// 	fmt.Println(route)
+// }
