+++ graphs/floydwarshall/floydwarshall.go
@@ -1,20 +1,19 @@
 // Floyd-Warshall algorithm
 // https://en.wikipedia.org/wiki/Floyd%E2%80%93Warshall_algorithm
 
-package graphs
+package floydwarshall
 
 import (
 	"math"
-	"fmt"
 )
 
-// Defining matrix to use 2d array easier
+// Matrix Defining matrix to use 2d array easier
 type Matrix [][]float64
 
 // Defining maximum value. If two vertices share this value, it means they are not connected
 var maxValue = math.Inf(1)
 
-// Returns all pair's shortest path using Floyd Warshall algorithm
+// FloydWarshall Returns all pair's shortest path using Floyd Warshall algorithm
 func FloydWarshall(graph Matrix) Matrix {
 	// If graph is empty or width != height, returns nil
 	if len(graph) == 0 || len(graph) != len(graph[0]) {
@@ -48,17 +47,17 @@ func FloydWarshall(graph Matrix) Matrix {
 	return result
 }
 
-func main() {
- 	var graph Matrix
- 	graph = Matrix{{0, maxValue, -2, maxValue},
- 		{4, 0, 3, maxValue},
- 		{maxValue, maxValue, 0, 2},
- 		{maxValue, -1, maxValue, 0}}
+// func main() {
+// 	var graph Matrix
+// 	graph = Matrix{{0, maxValue, -2, maxValue},
+// 		{4, 0, 3, maxValue},
+// 		{maxValue, maxValue, 0, 2},
+// 		{maxValue, -1, maxValue, 0}}
 
- 	result := FloydWarshall(graph)
+// 	result := FloydWarshall(graph)
 
- 	//Print result
- 	for i := 0; i < len(result); i++ {
- 		fmt.Printf("%4g\n", result[i])
- 	}
- }
+// 	//Print result
+// 	for i := 0; i < len(result); i++ {
+// 		fmt.Printf("%4g\n", result[i])
+// 	}
+// }
